% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_all.R
\name{collapse_all}
\alias{collapse_all}
\title{Assists in collapsing various designations of majors into more efficient major designations}
\usage{
collapse_all(a)
}
\arguments{
\item{a}{A file with AcadPlan with various designations of various majors}
}
\description{
Assists in collapsing various designations of majors into more efficient major designations
}
