% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_fs.R
\name{collapse_fs}
\alias{collapse_fs}
\title{Assists in collapsing various designation of family studies majors}
\usage{
collapse_fs(a)
}
\arguments{
\item{a}{A file containing AcadPlan with various designations of family studies majors}
}
\description{
Assists in collapsing various designation of family studies majors
}
