% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_health.R
\name{collapse_health}
\alias{collapse_health}
\title{Assists in collapsing categories for health majors}
\usage{
collapse_health(a)
}
\arguments{
\item{a}{A file with AcadPlan including various health major designations}
}
\description{
Assists in collapsing categories for health majors
}
