% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_tl.R
\name{collapse_tl}
\alias{collapse_tl}
\title{Assists with collapsing all kinds of designations of teaching and learning major}
\usage{
collapse_tl(a)
}
\arguments{
\item{a}{A file containing AcadPlan with various designations of teaching and learning majors}
}
\description{
Assists with collapsing all kinds of designations of teaching and learning major
}
