% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectXfer.R
\name{inspectXfer}
\alias{inspectXfer}
\title{a function to root out and delete spurious and duplicate transfer courses}
\usage{
inspectXfer(a = transfer_courses_file, b = transfer_hours_file)
}
\arguments{
\item{a}{transfer course file from UH peoplesoft, untouched}

\item{b}{transfer course hours file from UH peoplesoft, untouched}
}
\description{
a function to root out and delete spurious and duplicate transfer courses
}
