% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_zeros.R
\name{leading_zeros}
\alias{leading_zeros}
\title{Assists with creating leading zeros for either SSNs, IDs or TEAIDs}
\usage{
leading_zeros(e, f)
}
\arguments{
\item{e}{A file with either SSNs, IDs, or TEAIDs}

\item{f}{A choice that must be made of filling out either SSN, ID, or TEAID}
}
\description{
Assists with creating leading zeros for either SSNs, IDs or TEAIDs
}
