% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prin_pasl.R
\name{prinpasl}
\alias{prinpasl}
\title{a function to merge 268 and 368 scores to determine readiness for principal certification}
\usage{
prinpasl(a, b)
}
\arguments{
\item{a}{Test scores from Pearson, untouched}

\item{b}{Test from ETS, the PASL scores 368 for principal certification}
}
\description{
a function to merge 268 and 368 scores to determine readiness for principal certification
}
