% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_advanced.R
\name{recommend_advanced}
\alias{recommend_advanced}
\title{a function to help to determine if advanced students are ready for changes in status towards certification}
\usage{
recommend_advanced(a, b, d, e, f)
}
\arguments{
\item{a}{Test scores from Pearson, untouched}

\item{b}{Test from ETS, the PASL scores 368 for principal certification}

\item{d}{SSN numbers for students}

\item{e}{SSN and ETS ID numbers for students}

\item{f}{The academic year in question, in quotes; viz. "19-20"}
}
\description{
a function to help to determine if advanced students are ready for changes in status towards certification
}
