% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_initial.R
\name{recommend_initial}
\alias{recommend_initial}
\title{recommend_initial is a package to help make institutional effectiveness easier by tidying data to one row to one case and other actions}
\usage{
recommend_initial(a, b)
}
\arguments{
\item{a}{A file generated from Pearson downloads for students seeking initial certification}

\item{b}{The academic year in question, surrounded in quotes}
}
\value{
A data frame with decisions to recommend or not.
}
\description{
recommend_initial is a package to help make institutional effectiveness easier by tidying data to one row to one case and other actions
}
