% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_to_year.R
\name{term_to_year}
\alias{term_to_year}
\title{Converts a term into a semester and year}
\usage{
term_to_year(a)
}
\arguments{
\item{a}{a file with the "Term" variable duly noted and named in the dataframe}
}
\description{
Converts a term into a semester and year
}
