% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pasl.R
\name{update_pasl}
\alias{update_pasl}
\title{Assists in updating PASL scores from day to day.}
\usage{
update_pasl(a, b)
}
\arguments{
\item{a}{The new daily download from ETS}

\item{b}{old file with accurate SSN numbers}
}
\description{
Assists in updating PASL scores from day to day.
}
